Android additions to OkHttp
Additional files needed for OkHttp in Android (changes to
existing files follow in a subsequent commit). This changes
are reapplied from:
  - The additions (not modifications) from
    AOSP commit 36132b3cd2c814e057c775adc821ae0fb12aeb4c
  - Fix HttpURLConnection.setFollowRedirects()
    AOSP commit b701c07a57f6c8a871ed7a56860c6d0452bd740d
  - Refactor NetworkSecurityPolicy to be pluggable
    AOSP commit a8b361812f24441b359e797416e1251e2662f1c3
  - Add core-oj to the list of dependencies
    AOSP commit 71a49ecfabfc192b2bc962be777d46bc7f3ab11b
  - Use core-oj alongside core-libart wherever required.
    AOSP commit 04b2952d1e7276a2e53bfb1433bdda8a8523f910
  - Use hostname aware isCleartextTrafficPermitted
    AOSP commit 36e73c99548c96ca88c3d26bb82e0114acd46d9a
  - Force okhttp to continue to use v51 (1.7) .class files
    AOSP commit a314388155dea97b842fdaadda3385d3d7ed79a5

In addition to the contents of the above CLs, this CL also
adds methods:

   X509TrustManager trustManager(SSLSocketFactory)
   TrustRootIndex trustRootIndex(X509TrustManager)

to Platform.java to support new Certificate Pinning logic
introduced in OkHttp 2.7.5. This logic is not intended to
be involved in any API exposed by Android. We deliberately
don't expose it because Android has separate APIs and a
separate implementation of Certificate Pinning. Therefore
the new methods in Platform are stripped down to the
minimum required by OkHttp tests to pass. To help ensure
that these methods are not accidentally involved with
providing Android APIs, this CL adds functionality for
the Platform instance to be temporarily replaced during
tests.

Change-Id: Ib33e511c1837d44cb570e02aa47202d2b0260118