Revert "Fix regression in the HTTP request line"
This reverts commit 8e96772d82220b49aa27a4cbeca1578620fd9223.

This revert reintroduces a change the behavior of Http(s)URLConnection:
URLs with an empty path ("http://example.com") now get normalized to
a path with one empty element ("http://example.com/").

There was only one known case where the change in behavior caused
problems with a server, however that is no longer reproducable (perhaps
the server has been fixed). Therefore, it seems safe to introduce the
behavior change now.

The advantage of introducing this behavior change now is that we don't
need to carry this local patch. This makes it easier to update to later
versions of OkHttp in future.

Bug: 30107354
Change-Id: Ib505530274867375dabebb305e8005038d8f650d