DO NOT MERGE Relax N CTS about whether \t is allowed in header values
Android N's Http(s)URLConnection initially forbade '\t' occurring in
header values when it should be allowed; this was fixed by [1] which
will be part of a future Android release (a corresponding test was
updated in [2]). It's permitted for vendors shipping Android N to
apply the fixes [1]. Therefore, this CL applies a relaxed version of
the test changes from [1] to Android N CTS to allow either behavior.
CTS for APIs > 23 will enforce that \t is allowed.

[1] commit 1917ec9635dea723538ea000b67b105999e58710 in external/okhttp
[2] commit dcabb633be0f71e720a856e141867d5d23adff93 in libcore

Bug: 30799514

Test: cts-tradefed run cts -m CtsLibcoreOkHttpTestCases \
        -t com.squareup.okhttp.RequestTest
      against (a) N (b) a recent build that included [1-2].
Test: com.squareup.okhttp.internal.http.HeadersTest likewise

Change-Id: I069d5ad512f84228eaf85e83d75108d8bd492dfd